Encapsulate new client schemas and remove from broker/simple client defs
